Add support for expansion
This removes the notion of substruct and instead expands all
sub-modules and module types. It also expands packs and includes.

HTML generation now requires an index to do expansion against so
the command-line arguments and behaviour have been changed to match
that of link.

The expander requires an index from roots to units, so a root field
has been added to the unit information in indices.